[Coroutines] Move FlowAction.fromFlow to CoroutineAction.
