+++ formula-coroutines/src/main/java/com/instacart/formula/coroutines/LaunchCoroutineAction.kt
@@ -5,29 +5,29 @@ import com.instacart.formula.Cancelable
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 /**
- * A Formula [Action] that is created from a [Flow].
+ * Adapter which allows to run a suspend function as an [Action].
  */
-class FlowAction<Event>(
+class LaunchCoroutineAction<Result> @PublishedApi internal constructor(
     private val dispatcher: CoroutineDispatcher,
     private val key: Any?,
-    private val factory: suspend () -> Flow<Event>
-) : Action<Event> {
+    private val block: suspend () -> Result,
+) : Action<Result> {
 
     override fun key(): Any? = key
 
     @OptIn(DelicateCoroutinesApi::class)
-    override fun start(send: (Event) -> Unit): Cancelable {
+    override fun start(
+        send: (Result) -> Unit,
+    ): Cancelable {
         val job = GlobalScope.launch(start = CoroutineStart.UNDISPATCHED) {
             withContext(dispatcher) {
-                factory().collect { send(it) }
+                val result = block()
+                send(result)
             }
         }
         return Cancelable(job::cancel)
